requirements.txt
disnake~=2.7.0
APScheduler~=3.9.1.post1
python-dotenv~=0.21.1